requirements.txt
Django==1.11
djangorestframework==3.6.0
django-markdown==0.8.4
Markdown==2.6.11
psycopg2==2.7.5
pytz==2018.5
pytest-django==3.3.2
pytest-cov==2.5.1